Hack to download images from "productsup.io" #19496
This is ugly hack just to make sure client will not wait.
But I also prepared new endpoint where URL is encoded into
query parameter and not decoded automatically by URL router
so if used properly with "encodeURIComponent" on FE it will
work for any URL no matter if it has some encoded chars or not.